added comments and also moved activities to their respective sub folders.
