+++ app/src/main/java/org/samir/universitybazaar/Activity/Clubs/MySubscriptionsActivity.java
@@ -15,14 +15,14 @@ import org.samir.universitybazaar.Models.User;
 import org.samir.universitybazaar.R;
 
 import java.util.ArrayList;
-
+/**
+ * @author Samir Shrestha
+ * @description This class displays all the clubs that the user has subscribed to.
+ */
 public class MySubscriptionsActivity extends AppCompatActivity {
     private RecyclerView clubRecView;
     private ClubAdapter adapter;
-    /**
-     * @author Samir Shrestha
-     * @description This class displays all the clubs that the user has subscribed to.
-     */
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -31,10 +31,12 @@ public class MySubscriptionsActivity extends AppCompatActivity {
         clubRecView = findViewById(R.id.clubRecView);
         adapter = new ClubAdapter(this);
 
-        ClubDAO db = new ClubDAO(this);
-        UserSession userSession = new UserSession(this);
+        ClubDAO db = new ClubDAO(this);//handle for database access
 
+        //get current logged in user
+        UserSession userSession = new UserSession(this);
         User user = userSession.isUserLoggedIn();
+
         if(user != null){ //user is logged in.
             ArrayList<Club> clubs = new ArrayList<>();
             clubs = db.getSubscribedClubsByMemberId(user.getMemberId()); //get all the clubs whose memberId matches the logged in users memberId in the subscriptions list view.
