+++ app/src/main/java/org/samir/universitybazaar/Adapter/CommentAdapter.java
@@ -11,10 +11,8 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
-import org.samir.universitybazaar.Activity.CommentReplyActivity;
-import org.samir.universitybazaar.Activity.Profile.EditProfileActivity;
+import org.samir.universitybazaar.Activity.Posts.CommentReplyActivity;
 import org.samir.universitybazaar.Activity.Posts.PostActivity;
-import org.samir.universitybazaar.Activity.Profile.ViewProfileActivity;
 import org.samir.universitybazaar.Database.CommentDAO;
 import org.samir.universitybazaar.Database.DatabaseHelper;
 import org.samir.universitybazaar.Database.UserSession;
