+++ app/src/main/res/layout/activity_comment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity.CommentActivity">
+    tools:context=".Activity.Posts.CommentActivity">
     tools:context=".Activity.Clubs.CommentOnNoticeActivity">
 
     <TextView
