+++ app/src/main/res/layout/activity_comment_reply.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity.CommentReplyActivity">
+    tools:context=".Activity.Posts.CommentReplyActivity">
 
     <TextView
         android:id="@+id/txtReply"
