+++ app/src/main/res/layout/activity_edit_post.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Activity.EditPostActivity">
+    tools:context=".Activity.Posts.EditPostActivity">
 
     <TextView
         android:id="@+id/txtUpdatePost"
