+++ app/src/main/res/layout/fragment_home.xml
@@ -24,6 +24,15 @@
             android:textColor="#3F51B5" />
 
         <TextView
+            android:id="@+id/txtMessages"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:textSize="20sp"
+            android:layout_centerInParent="true"
+            android:text="MESSAGES(0)"
+            android:textColor="#3F51B5" />
+
+        <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="LOGOUT"
